Line actions as state container
