+++ react/components/EXPERIMENTAL_Table/contexts.ts
@@ -1,9 +1,7 @@
 import { useContext, createContext } from 'react'
 import { OUT_OF_SCOPE_COMPOSITES_ERROR } from './errors'
-import { Bulk } from './hooks/useTableBulkActions'
 
 export const TableContext = createContext<Partial<TableState & TableProps>>({})
-export const BulkContext = createContext<Partial<Bulk>>({})
 
 export const useTableContext = () => {
   const context = useContext(TableContext)
@@ -12,5 +10,3 @@ export const useTableContext = () => {
   }
   return context
 }
-
-export const useBulkContext = () => useContext(BulkContext)
